sui_display::v2

Trait Store

pub trait Store {
    // Required method
    fn object<'life0, 'async_trait>(
        &'life0 self,
        id: AccountAddress,
    ) -> Pin<Box<dyn Future<Output = Result<Option<OwnedSlice>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Dynamically load objects by their ID, returning the object’s owned data.

The Store trait is responsible only for fetching object data – lifetime management and caching are handled by the Interpreter.

Required Methods

fn object<'life0, 'async_trait>( &'life0 self, id: AccountAddress, ) -> Pin<Box<dyn Future<Output = Result<Option<OwnedSlice>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,